Add net and row file generation
This commit adds the ability to generate net and row files.

The net file is generated using the number of lanes and the road length.
The row file is generated using the max speed, the number of vehicles per hour, the begin time, the end time and the edge.

The net and row files are generated in the simulation folder.